done2

+++ src/Main.java
@@ -84,8 +84,8 @@ public class Main {
 
     private static void task9() {
         System.out.println("Задача 9");
-        int money = 29000;
-        int total = 0;
+        double money = 29000;
+        double total = 0;
         for (int i = 0; i < 12; i++) {
             total = total + total / 100;
             total = total + money;
